requirements.txt
Django==1.11.7
django-celery==3.2.2
django-cors-headers==2.2.0
djangorestframework==3.7.7
pandas==0.22.0
pymongo==3.5.1
PyMySQL==0.7.11
SQLAlchemy==1.1.15

